requirements.txt
streamlit
torch
torchvision
torchaudio
opencv-python
Pillow
numpy
pandas
tqdm
matplotlib
seaborn
python-dateutil
joblib
scikit-learn
albumentations
jupyter
notebook
protobuf